Fix compiler warnings in chromeos_ime.cc and monitor_language.cc.
I wonder why ibus_chromeos_panel_service_new() was working before the change...

TEST=built with scons --f SConstruct.chromiumos

Review URL: http://codereview.chromium.org/543186

+++ monitor_language.cc
@@ -33,7 +33,7 @@
 
 namespace {
 
-const size_t kTestCount = 5;
+const int kTestCount = 5;
 chromeos::LanguageStatusConnection* global_connection = NULL;
 
 void DumpProperties(const chromeos::ImePropertyList& prop_list) {
